Revert "MM-7 feature/#7: 로그인 기능 개발"
