+++ .gitignore
@@ -69,4 +69,3 @@ yarn-error.log
 yarn.lock
 .yarn
 yarn
-package.json
