+++ package.json
@@ -10,8 +10,6 @@
 		"test": "jest"
 	},
 	"dependencies": {
-		"@emotion/native": "^11.11.0",
-		"@emotion/react": "^11.11.4",
 		"@react-native-masked-view/masked-view": "^0.3.1",
 		"@react-navigation/bottom-tabs": "^6.5.20",
 		"@react-navigation/native": "^6.1.17",
