+++ src/screens/auth/LoginScreen.tsx
@@ -1,0 +1,15 @@
+import { StyleSheet, Text, View } from 'react-native';
+
+interface LoginScreenProps {}
+
+function LoginScreen({}: LoginScreenProps) {
+	return (
+		<View>
+			<Text>로그인 스크린</Text>
+		</View>
+	);
+}
+
+const styles = StyleSheet.create({});
+
+export default LoginScreen;
