+++ src/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './header';
 export * from './encryptStorage';
-export * from './validate';
