--- src/utils/validate.ts
@@ -1,27 +1,0 @@
-type UserInformation = {
-	email: string;
-	password: string;
-};
-
-function validateUser(values: UserInformation) {
-	const errors = {
-		email: '',
-		password: '',
-	};
-
-	if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(values.email)) {
-		errors.email = '올바른 이메일 형식이 아닙니다.';
-	}
-
-	if (!(values.password.length >= 8 && values.password.length < 20)) {
-		errors.password = '비밀번호는 8~20자 사이로 입력해주세요.';
-	}
-
-	return errors;
-}
-
-function validateLogin(values: UserInformation) {
-	return validateUser(values);
-}
-
-export { validateLogin };
