+++ tsconfig.json
@@ -5,5 +5,5 @@
     "paths": {
       "@/*": ["src/*"]
     }
-  },
+  }
 }
